Move rizzle.ts to util
